Struct uic_gms::mqtt_handler::MqttGroupHandler

pub struct MqttGroupHandler<T: MqttClientTrait> {
    client: T,
    cache: Cache,
    reported_group_list_matcher: TopicMatcherType,
    reported_names_matcher: TopicMatcherType,
    supported_commands: TopicMatcherType,
}

MqttGroupHandler handles subscription messages related to UIC groups. It subscribes to listen to the correct topics and is to be passed to the mosquitto client

Fields
client: Tcache: Cache

Lookup cache for groups and unids

reported_group_list_matcher: TopicMatcherTypereported_names_matcher: TopicMatcherTypesupported_commands: TopicMatcherType
Implementations

impl<T: MqttClientTrait> MqttGroupHandler<T>

pub fn new(client: T) -> Result<Self, sl_status_t>

crate new handler to be passed into a MqttClient. The native mqtt client is passed as reference counted. Reference counted objects can be shared and borrowed over multiple instances. Acquirring mutablity is only possible when the ref count == 1. since this rc is shared between MqttClient and MqttGroupHandler, reference count can never be == 1. hence mutable cannot be accuired for mqtt client.

fn process_message(&mut self, msg: MosqMessage) -> Result<(), String>

fn on_group_list_reported(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>

process a ucl/by-unid/+/+/Groups/Attributes/GroupList/Reported topic stores a list of groups for an unid and publishes its endpoints on the ucl/by-group//NodeList/ expects endpoint to be prefixed with ep. e.g. ep0

fn publish_commands_by_cluster(
&mut self,
group_id: &u16,
supported_commands: BTreeMap<String, HashSet<String>>,
is_unretain: bool
) -> Result<(), String>

fn on_name_reported(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>

handler for topic ucl/by-unid/+/+/Groups/Attributes/+/Name/Reported. overwrites name for group id.

fn send_add_group_if_required(
&self,
group_id: u16,
group_name: &str
) -> Result<(), String>

add_group is only send when an update of an groupname is reported

fn send_group_name(&self, group_id: u16, group_name: &str) -> Result<(), String>

fn on_supported_commands(
&mut self,
topic: &str,
json: JsonValue
) -> Result<(), String>